BUGFIX: correct conversion factor in actual_et__fao56__two_stage: 0.3048 => 0.328 as default minimum rooting depth.

+++ src/crop_coefficients__fao56.F90
@@ -480,7 +480,7 @@ end function update_crop_coefficient_date_as_threshold
   real (kind=c_float)                :: fKcb
 
   ! [ LOCALS ]
-  real (kind=c_float) :: fFrac
+  real (kind=c_double) :: fFrac
 
   ! define shorthand variable names for remainder of function
   associate ( GDD_ini_l => GROWTH_STAGE_GDD( GDD_INI, iLanduseIndex ),         &
